Clone restClientOptions rather than mutating parameter
The same instance of RestClientOptions is passed to the events API client and the cdn API client, both mutate the base url, which was causing the events requests to be sent to the cdn url.
Create DvcRestClientOptions class that extends RestClientOptions and adds a Clone method. This allows us to clone the rest client options before mutating.